Organizes Battles and Foes
Adds /Battle and /Foe
